Mod split util + autograde
